First generalisation.
Included functions f(x) and f'(x). It calculates the root by the general expression.

x0 = xi - f(x)/f'(x)

where, x0 iteratively approaches to the value of the root.
and, xi is the initial guess value which is initiated to x0, to be used in the next loop.